Добавил реализацию репозиториев, перевел id в Long

+++ src/main/resources/db/migration/V1__init_migration.sql
@@ -1,13 +1,13 @@
 CREATE TABLE weather_viewer.users (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     login VARCHAR(100) NOT NULL UNIQUE,
     password VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE weather_viewer.locations (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(100) NOT NULL,
-    user_id INT NOT NULL,
+    user_id BIGINT NOT NULL,
     latitude NUMERIC(9, 6) CHECK (latitude BETWEEN -90 AND 90),
     longitude NUMERIC(10, 6) CHECK (longitude BETWEEN -180 AND 180),
     CONSTRAINT fk_user FOREIGN KEY (user_id)
@@ -16,7 +16,7 @@ CREATE TABLE weather_viewer.locations (
 
 CREATE TABLE weather_viewer.sessions (
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    user_id INT NOT NULL,
+    user_id BIGINT NOT NULL,
     expires_at TIMESTAMP,
     CONSTRAINT fk_user FOREIGN KEY (user_id)
         REFERENCES weather_viewer.users (id) ON DELETE CASCADE
